+++ simulate/simulate.cc
@@ -739,18 +739,12 @@ void MakePhysicsSection(mj::Simulate* sim) {
   for (int i=0; i<mjNDISABLE; i++) {
     mju::strcpy_arr(defFlag[0].name, mjDISABLESTRING[i]);
     defFlag[0].pdata = sim->disable + i;
-    if ((1 << i) == mjDSBL_AUTORESETNAN) {
-      defFlag[0].state = 0;
-    } else {
-      defFlag[0].state = 2;
-    }
     mjui_add(&sim->ui0, defFlag);
   }
   mjui_add(&sim->ui0, defEnableFlags);
   for (int i=0; i<mjNENABLE; i++) {
     mju::strcpy_arr(defFlag[0].name, mjENABLESTRING[i]);
     defFlag[0].pdata = sim->enable + i;
-    defFlag[0].state = 2;
     mjui_add(&sim->ui0, defFlag);
   }
   // add contact override
@@ -1956,6 +1950,7 @@ void Simulate::Sync() {
   if (pending_.reset) {
     mj_resetData(m_, d_);
     mj_forward(m_, d_);
+    load_error[0] = '\0';
     update_profiler = true;
     update_sensor = true;
     scrub_index = 0;