+++ unity/Runtime/Bindings/MjBindings.cs
@@ -157,7 +157,7 @@ public enum mjtDisableBit : int{
   mjDSBL_SENSOR = 4096,
   mjDSBL_MIDPHASE = 8192,
   mjDSBL_EULERDAMP = 16384,
-  mjDSBL_AUTORESETNAN = 32768,
+  mjDSBL_AUTORESET = 32768,
   mjNDISABLE = 16,
 }
 public enum mjtEnableBit : int{