+++ doc/changelog.rst
@@ -8,7 +8,8 @@ Upcoming version (not yet released)
 General
 ^^^^^^^
 1. Renamed ``mjModel.tex_rbg`` to ``mjModel.tex_data``.
-2. Added a new disable flag ``mjDSBL_AUTORESETNAN`` to disable automatic reset when NaNs or infinities are produced.
+2. Added a new :ref:`autoreset<option-flag-autoreset>` flag to disable automatic reset when NaNs or infinities are
+   detected.
 3. Added sub-elements to the MJCF :ref:`material<asset-material>` element, to allow specification of multiple textures
    for rendering (e.g., :ref:`occlusion-roughness-metallic<material-orm>`). Note that the MuJoCo renderer doesn't
    support these new features, and they are made available for use with external renderers.